# docs/openapi.yaml
openapi: 3.0.0
info:
  title: E-Commerce REST API
  description: |
    A complete REST API for an online shopping platform.

  version: 1.0.0
  contact:
    name: API Support
    email: apisupportemail@ecommerce.com
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
  - url: https://your-app.onrender.com/api
    description: Production Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token from login
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: user@example.com
        role:
          type: string
          enum: [USER, ADMIN]
          example: USER
...
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Laptop"
        description:
          type: string
          nullable: true
          example: "High-performance laptop with 16GB RAM"
        price:
          type: number
          format: float
          example: 999.99
        stock:
          type: integer
          example: 10
        categoryId:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - name
        - price
        - categoryId
    ProductWithCategory:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            category:
              $ref: '#/components/schemas/Category'
    CreateProductInput:
      type: object
      required:
        - name
        - price
        - categoryId
      properties:
        name:
          type: string
          minLength: 2
          example: "Smartphone"
        description:
          type: string
          example: "Latest model smartphone"
        price:
          type: number
          minimum: 0.01
          example: 699.99
        categoryId:
          type: integer
          minimum: 1
          example: 1
        stock:
          type: integer
          minimum: 0
          default: 0
          example: 25
...
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "Electronics"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CategoryWithProducts:
      allOf:
        - $ref: '#/components/schemas/Category'
        - type: object
          properties:
            products:
              type: array
              items:
                $ref: '#/components/schemas/Product'
    CreateCategoryInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          example: "Home & Garden"
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        price:
          type: number
          format: float
        product:
          $ref: '#/components/schemas/Product'
    Order:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          example: 2
        status:
          type: string
          enum: [PENDING, COMPLETED, CANCELLED]
          example: "PENDING"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        user:
          $ref: '#/components/schemas/UserResponse'
    CreateOrderItemInput:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
          minimum: 1
          example: 1
        quantity:
          type: integer
          minimum: 1
          example: 2
    CreateOrderInput:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateOrderItemInput'
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message describing what went wrong"
    ValidationError:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
          example: ["Name is required", "Price must be a positive number"]
paths:
...
  /products:
    get:
      summary: Get all products
      description: Returns a list of all products with optional filtering and pagination.
      tags: [Products]
      security: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by product name or description
        - name: categoryId
          in: query
          schema:
            type: integer
            minimum: 1
          description: Filter by category ID
        - name: minPrice
          in: query
          schema:
            type: number
            minimum: 0
          description: Minimum price filter
        - name: maxPrice
          in: query
          schema:
            type: number
            minimum: 0
          description: Maximum price filter
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [id, name, price, stock, createdAt]
            default: id
          description: Field to sort by
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort order
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Pagination offset
      responses:
        200:
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductWithCategory'
        500:
          description: Internal server error

    post:
      summary: Create a new product
      description: Creates a new product. **Admin only**
      tags: [Products]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductInput'
      responses:
        201:
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        401:
          description: Not authenticated - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Product name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /categories:
    get:
      summary: Get all categories
      description: Returns a list of all categories with their products.
      tags: [Categories]
      security: []
      responses:
        200:
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryWithProducts'

    post:
      summary: Create a new category
      description: Creates a new category. **Admin only**
      tags: [Categories]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryInput'
      responses:
        201:
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        400:
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        401:
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        403:
          description: Forbidden - Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        409:
          description: Category name already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /orders:
    get:
      summary: Get user's orders
      description: |
        Returns all orders for the authenticated user.
        - **Regular users see only their own orders
        - **Admins see all orders
      tags: [Orders]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        401:
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new order
      description: Creates a new order for the authenticated user. Automatically checks stock availability and updates inventory.
      tags: [Orders]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderInput'
      responses:
        201:
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...